LPS-76927 Move new upgrade process to 7_1_x
