+++ portal-impl/src/com/liferay/portal/upgrade/v7_1_x/util/CounterTable.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.portal.upgrade.v7_1_0.util;
+package com.liferay.portal.upgrade.v7_1_x.util;
 
 import java.sql.Types;
